+++ src/main/java/dk/kb/netarchivesuite/solrwayback/service/SolrWaybackMementoAPI.java
@@ -1,6 +1,9 @@
 package dk.kb.netarchivesuite.solrwayback.service;
 
 import dk.kb.netarchivesuite.solrwayback.memento.DatetimeNegotiation;
+import dk.kb.netarchivesuite.solrwayback.memento.TimeMap;
+import dk.kb.netarchivesuite.solrwayback.service.exception.NotFoundServiceException;
+import dk.kb.netarchivesuite.solrwayback.service.exception.SolrWaybackServiceException;
 import dk.kb.netarchivesuite.solrwayback.util.PathResolver;
 import org.apache.http.client.utils.DateUtils;
 import org.slf4j.Logger;
@@ -20,7 +23,6 @@ import java.net.URI;
 import java.net.URISyntaxException;
 import java.util.Date;
 
-import static dk.kb.netarchivesuite.solrwayback.memento.TimeMap.getTimeMap;
 import static dk.kb.netarchivesuite.solrwayback.util.DateUtils.validateTimestamp;
 
 /**
@@ -81,13 +83,27 @@ public class SolrWaybackMementoAPI {
         mimeTypeForResponse = getMimeTypeForResponse(type, mimeTypeForResponse);
 
         URI uri =  PathResolver.mementoAPIResolver("/timemap/" + type + "/", uriInfo, url);
-        StreamingOutput timemap = getTimeMap(uri, type, null);
+        TimeMap timeMap = new TimeMap();
+        StreamingOutput timemap = timeMap.getTimeMap(uri, type, null);
         String fileType = fileEndingFromAcceptHeader(mimeTypeForResponse);
 
         return Response.ok().type(mimeTypeForResponse)
                 .entity(timemap)
                 .header(HttpHeaders.CONTENT_DISPOSITION, "inline") // Ensure inline displa
                 .build();
+    }
+
+    /**
+     * The paged timemap is not implemented for the json timemap, therefore trying to access a paged version throws an exception.
+     */
+    @GET
+    @Path("timemap/{page:\\d+}/json/{url:.+}")
+    public Response timeMapPagedJson(@Context UriInfo uriInfo, @Context HttpServletRequest httpRequest, @PathParam("url") String url, @PathParam("page") String page)
+            throws NotFoundServiceException {
+
+        String message = "Endpoint timemap/json does not support pagination. Either access the full timemap at /services/memento/timemap/json/" + url + " or request a paged " +
+                "timemap in the link or spec-json format at /services/memento/timemap/" + page + "/spec/" + url;
+        throw new NotFoundServiceException(message);
     }
 
     @GET
@@ -102,12 +118,13 @@ public class SolrWaybackMementoAPI {
         mimeTypeForResponse = getMimeTypeForResponse(type, mimeTypeForResponse);
 
         URI uri =  PathResolver.mementoAPIResolver("/timemap/" + page + "/" + type + "/", uriInfo, url);
-        StreamingOutput timemap = getTimeMap(uri, type, Integer.valueOf(page));
+        TimeMap timeMap = new TimeMap();
+        StreamingOutput output = timeMap.getTimeMap(uri, type, Integer.valueOf(page));
         String fileType = fileEndingFromAcceptHeader(mimeTypeForResponse);
 
         // TODO: Fresh eyes on http headers for timemap
         return Response.ok().type(mimeTypeForResponse)
-                .entity(timemap)
+                .entity(output)
                 .header(HttpHeaders.CONTENT_DISPOSITION, "inline") // Ensure inline displa
                 //.header(HttpHeaders.CONTENT_DISPOSITION, "attachment ; filename = \"timemap"+ fileType + "\"")
                 .build();