+++ CHANGES.md
@@ -4,10 +4,11 @@ UNRELEASED
 -----
 
 
+
+
 5.2.1
 -----
 * Fixed memento null-pointer.
-
 
 5.2.0
 -----